backend/services/search-indexing/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
redis==5.0.1

# Elasticsearch
elasticsearch==8.11.0
aiohttp==3.9.1

# Event handling
kafka-python==2.0.2

# Security & Auth
python-jose[cryptography]==3.3.0
httpx==0.25.2

# Utilities
structlog==23.2.0